+++ src/specification/timestamp.rs
@@ -3,7 +3,7 @@ use std::{
     str::FromStr,
 };
 
-use chrono::{DateTime, FixedOffset, NaiveDateTime, TimeZone};
+use chrono::{DateTime, FixedOffset, NaiveDateTime, Offset, TimeZone, Utc};
 use thiserror::Error;
 
 /// Represents a timestamp in a context diff file.
@@ -30,15 +30,16 @@ impl FromStr for Timestamp {
     /// GNU: 2002-02-21 23:30:39.942229878 -0800 (Note that the fractional second part can be omitted)
     /// POSIX: Thu Feb 21 23:30:39 2002
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        if s.is_empty() {
+        // Get first character if available, return empty error otherwise
+        let Some(first_char) = s.chars().next() else {
             return Err(TimestampParseError::TimestampIsEmpty);
-        }
+        };
 
         // Check if the first character is a-z, which would indicate a POSIX timestamp
-        if s.chars().next().expect("Expected a first character").is_ascii_alphabetic() {
+        if first_char.is_ascii_alphabetic() {
             let naive = NaiveDateTime::parse_from_str(s, "%a %b %e %H:%M:%S %Y").map_err(TimestampParseError::ChronoParseError)?;
             return Ok(Self {
-                value: FixedOffset::east_opt(0).expect("Expected 0 to be a valid offset").from_utc_datetime(&naive),
+                value: Utc.fix().from_utc_datetime(&naive),
             });
         }