+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added
 - Test cases for the full parsing and translation from context diffs to unified diffs.
 
+### Changes
+- The parser now does not use `expect` anymore, eliminating unexpected panics.
+
 ### Fixed
 - Fix wrong expected hunk length for hunks of files that don't exist (new or removed files).
 - Fix invalid hunks where start and end line or equal being accepted.